# Make opt_eq and opt_neq insns leaf