@k0kubun 2019-08-04
More information related to https://bugs.ruby-lang.org/issues/15589:

```rb
require 'benchmark/ips'

Numeric.class_eval do
  def ruby_zero?
    self == 0
  end
end

Benchmark.ips do |x|
  x.report('0.ruby_zero?') { 0.ruby_zero? }
  x.report('1.ruby_zero?') { 1.ruby_zero? }
  x.report('0.zero?') { 0.zero? }
  x.report('1.zero?') { 1.zero? }
  x.report('0 == 0') { 0 == 0 }
  x.report('1 == 0') { 1 == 0 }
  x.compare!
end
```

## before
```
              1 == 0: 46145043.2 i/s
              0 == 0: 45733750.0 i/s - same-ish: difference falls within error
        0.ruby_zero?: 36376164.9 i/s - 1.27x  slower
        1.ruby_zero?: 36208189.9 i/s - 1.27x  slower
             1.zero?: 31974176.6 i/s - 1.44x  slower
             0.zero?: 31394964.5 i/s - 1.47x  slower
```

## after
```
              0 == 0: 47318096.6 i/s
              1 == 0: 47202782.4 i/s - same-ish: difference falls within error
        1.ruby_zero?: 44869590.8 i/s - same-ish: difference falls within error
        0.ruby_zero?: 44515046.6 i/s - same-ish: difference falls within error
             0.zero?: 33165785.7 i/s - 1.43x  slower
             1.zero?: 33133627.8 i/s - 1.43x  slower
```

In JIT, the performance difference between `== 0` and `.zero?` (implemented in Ruby) will become very short by this patch. If it has this level of small difference, I think it's fine to close [Bug #15589].

In the next patch following this PR will allow overriding `#zero?` with Ruby only when `--jit` is passed.