@ko1 2019-08-04
The following description is deep background by @k0kubun in Japanese. I put it here to note the background.
If someone need English description to understand this PR, we'll translate it.

```
1. Numeric#zero? を self == 0 と定義した時、 a == 0 と a.zero? の性能の差はvm_push_frameが支配的である
2. メソッドA のinsnがleaveを除き全てleafである時、leaveのCHECK_INTSを外せば メソッドA を呼び出す際vm_push_frameをしなくても問題ない
3. 変更前の opt_eq がleafにならない要因は、レシーバがStringだが引数がStringでない場合に rb_equal (内部で rb_funcall する) を使うためである

という仮定の上で、

何をどうした: "レシーバがStringだが引数がStringでない場合" に opt_equal_fallback にフォールバックするようにし、フォールバック時にcall cacheにrb_obj_equalがキャッシュされてない場合は Qundef が返る (= Qundefな時は opt_send_without_block に飛ぶので、opt_eqにいる間にrb_funcallが呼ばれない) ようにした。

なんのために: Ruby で Numeric#zero? を定義し直すだけで vm_push_frame なしのインライン化が（MJITで）できるようにするため。
```